Implement `q_remove_head` for head remove
The `q_remove_head` function removes the first element of the queue and
returns it. In addition, The function does not free the memory of the
removed element. If the queue is empty, the function returns `NULL`. If
the `sp` parameter is not `NULL`, the function creates a copy of the
removed element and stores it in the memory pointed by `sp`.