Implement `q_delete_mid` to delete middle element
The `q_delete_mid` function should delete the middle element of the
queue. If the queue is empty, the function will return `false`. The
method which finds the middle element is to search both clockwise and
counter-clockwise from the head and the tail of the queue. If the
pointers match or nearby, then the middle element is found.